requirements.txt
aniso8601==8.0.0
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
blueprint==3.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
coverage==5.1
ez-setup==0.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
flask-talisman==0.7.0
Flask-WTF==0.14.3
gunicorn==20.0.4
h5py==2.9.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.19.1
mysqlclient==1.4.6
numpy==1.15.4
pandas==0.23.4
pbr==5.1.1
Pillow==5.4.0
psycopg2==2.8.5
pyarcade==0.0.1
pycparser==2.19
pylint==2.4.4
pymongo==3.10.1
pyotp==2.3.0
python-dateutil==2.7.5
python-dotenv==0.12.0
pytz==2018.7
PyYAML==3.13
qrcode==6.1
requests==2.23.0
scikit-learn==0.20.2
scipy==1.2.0
sequential==1.0.0
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.16
stevedore==1.30.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.2.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
